♻️ refactor(cache): redesign the cache api to more simpler design & update the module structure (#623)
